Add frequency decay
This is the simplest possible frequency decay: simply multiply every
non-chosen item by a constant factor *decay*.  This is 1 by
default (i.e. don't use a frequency decay at all) for backwards
compatibility.

+++ src/Core/Util.hs
@@ -24,7 +24,7 @@ import qualified System.Posix.FilePath as BS -- used for ByteString version of <
 
 -- | Type for an Map that describes all of the executables with their
 -- ratings.
-type Items = Map ByteString Int
+type Items = Map ByteString Double
 
 -- | Type for helping to decide how to open something.
 data OpenIn = Term ShowBS | Open ShowBS
